Changed the env check if we are standalonifying to use the new: ENV['RUBYCOCOA_STANDALONEIFYING?']
This will be committed to RubyCocoa in a couple of minutes, so this needs a fresh checkout/update.

Standalonify is also updated to skip any files that already are inside the bundle resources somwhere.

Commented a "require 'rubygems'" statement in lib/server.rb because we only use webrick atm. This means that rubygems will not be bundled and thus the size will be cut down by 1MB.

git-svn-id: https://fngtps.com/svn/kari/trunk/src@128 7a99d2f0-d533-0410-96fe-f25cb2755c1e